Extract setting network parameters and allow setting constant values

+++ mldr/envs/ns3_ai_structures.py
@@ -17,7 +17,7 @@ class Env(Structure):
 class Act(Structure):
     _pack_ = 1
     _fields_ = [
-        ('cw', c_uint8),
+        ('cw', c_int),
         ('rts_cts', c_bool),
         ('ampdu', c_bool),
         ('end_warmup', c_bool)
